pyproject.toml
[project]
name = "kpop-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
kpop-agent = "kpop_agent:main"
start-agent="kpop_agent.helloAI:run_agent"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
